requirements.txt
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
HTMLParser==0.0.2
idna==2.6
matplotlib==2.1.0
nltk==3.2.5
numpy==1.13.3
oauthlib==2.0.6
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
